Clean up example5 Generate chapter 5's notebook
